requirements-dev.txt
Flask-DebugToolbar
ipdb
ipython
pytest
